+++ crypto-common/CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.2.1 (UNRELEASED)
+### Changed
+- `BlockSizeUser::BlockSize` is no longer bounded by `BlockSizes` ([#2309])
+
+### Removed
+- `BlockSizes` trait ([#2309])
+
+[#2309]: https://github.com/RustCrypto/traits/pull/2309
+
 ## 0.2.0 (2026-02-04)
 ### Added
 - Sealed `BlockSizes` trait implemented for types from `U1` to `U255` ([#1172])